pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillscan"
version = "0.1.0"
description = "Standalone AI skill security scanner"
readme = "README.md"
authors = [{name = "SkillScan"}]
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.1",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.4",
  "mypy>=1.10.0"
]

[project.scripts]
skillscan = "skillscan.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
skillscan = ["data/policies/*.yaml", "data/intel/*.json", "data/rules/*.yaml"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 110
target-version = "py312"
exclude = ["tests/adversarial/cases"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
